Small changes to make running dlv interactively easier.
Mainly ensuring single call to nameList(), and not changing behavior if
DNS resolution process is slow.

In combination with go 1.20.1 fb06acf511, does single nameList() call
and gets all the way to the panic in Ping even if sleep for a minute in
nameList().

Some changes may not be needed - they are:
- check for localhost in go
- remove query timeout
- remove other call to Ping()
- single loop iteration in WaitMySqlWorks

Also:
- remove rr stuff in build
- copy in go nss file
- add dlv init file that sets initial wp